refactor(vnext): relocate logger + paths (#623 step 1)
Pure relocate, no logic change. First step in the bottom-up core/ → vnext/
rebuild per .claude/prd/core-cleanup-vnext.md.

- src/core/logger.{ts,test.ts} → src/vnext/logging/
- src/core/paths.{ts,test.ts}  → src/vnext/  (top-level file)
- update 49 import sites
- sync header comments
- PRD updated: logger is cross-cutting (own folder, peer to routing/transport/...);
  paths is project-specific top-level; utils/ deferred until a real generic util lands.
