+++ src/automation/cron-job.ts
@@ -2,7 +2,7 @@
 // Manages cron-based scheduled tasks for agents and channels.
 
 import { Cron } from "croner";
-import { createLogger } from "../core/logger.js";
+import { createLogger } from "../vnext/logging/logger.js";
 
 const log = createLogger("cron");
