+++ src/automation/heartbeat.test.ts
@@ -29,7 +29,7 @@ vi.mock("node:fs/promises", () => ({
 }));
 
 // Suppress log output in tests
-vi.mock("../core/logger.js", () => ({
+vi.mock("../vnext/logging/logger.js", () => ({
   createLogger: () => ({
     debug: vi.fn(),
     info: vi.fn(),
