+++ src/commands/slash-commands.ts
@@ -4,7 +4,7 @@
 import type { SessionStore } from "../core/types.js";
 import type { AgentServiceCache } from "../core/pi-mono.js";
 import type { DefaultAgentManager } from "../core/agent-manager.js";
-import { createLogger } from "../core/logger.js";
+import { createLogger } from "../vnext/logging/logger.js";
 import { failureTracker } from "../agents/failure-tracker.js";
 
 const log = createLogger("commands");
