+++ src/core/agent-run.ts
@@ -5,7 +5,7 @@
 import type { AgentRuntime } from "../agents/runtime.js";
 import { agentEventBus } from "./agent-event-bus.js";
 import { userMessage, assistantMessage, getAgentEndMeta, getUsage } from "./messages.js";
-import type { Logger } from "./logger.js";
+import type { Logger } from "../vnext/logging/logger.js";
 import type { UsageTracker } from "./usage-tracker.js";
 import type { HookRegistry } from "../plugins/hooks.js";
 import { runWithMessageContext } from "../transport/context.js";
