+++ src/plugins/discord/discord.ts
@@ -22,7 +22,7 @@ import type { DefaultAgentManager } from "../../core/agent-manager.js";
 import { userMessage as mkUserMsg, userMessageWithImages as mkUserMsgWithImages } from "../../core/messages.js";
 import type { ContextConfigFile } from "../../core/config.js";
 import { shouldRespondToMessage } from "../../core/mention.js";
-import { loggers } from "../../core/logger.js";
+import { loggers } from "../../vnext/logging/logger.js";
 import { ThreadBindingManager } from "./thread-bindings.js";
 import { consumeRootRun, cancelRunBySessionId, isRootRunActive } from "../../core/agent-run.js";
 import { runWithDiscordSubagentStream, type DiscordSubagentStreamContext } from "./subagent-stream-context.js";
